Update listener code for crate menus
