+++ paper/run/server.properties
@@ -1,5 +1,5 @@
 #Minecraft server properties
-#Tue Feb 06 03:13:15 EST 2024
+#Sat Feb 17 13:54:46 EST 2024
 enable-jmx-monitoring=false
 rcon.port=25575
 level-seed=
